got lightning working, using unet-mri model from pytorch got loss of .08, still running

+++ preprocess.py
@@ -57,7 +57,7 @@ def get_windows(img_f, mask):
                 r_chunks = split(r)
                 mask_chunks = split(mask_check)
                 for i in range(4):
-                    samples.append((torch.from_numpy(r_chunks[i]), torch.from_numpy(mask_chunks[i])))
+                    samples.append((torch.from_numpy(r_chunks[i]).float(), torch.from_numpy(mask_chunks[i]).float()))
                 # also can probably convert to tensors here as well
                # samples.append((torch.from_numpy(r),torch.from_numpy(mask_check)))
             else:
@@ -80,7 +80,7 @@ def get_vegetation_index_windows(img_f, mask):
                 mask_chunks = split(mask_check)
                 # the split function return 4 separate quadrants from the original window
                 for i in range(4):
-                    samples.append((torch.from_numpy(chunks[i]), torch.from_numpy(mask_chunks[i])))
+                    samples.append((torch.from_numpy(chunks[i]).float(), torch.from_numpy(mask_chunks[i]).float()))
                 # also can probably convert to tensors here as well,
                # samples.append((torch.from_numpy(r),torch.from_numpy(mask_check)))
             else:
